requirements.txt
streamlit==1.32.0
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.2.0
plotly==5.19.0
anthropic==0.18.1
APScheduler==3.10.4
python-dateutil==2.8.2
pytz==2024.1
# New dependencies for enhanced features
email-validator==2.1.0
openpyxl==3.1.2